Moved rtm io ui and property definitions
