+++ Arma3ObjectBuilder/ui/__init__.py
@@ -1,6 +1,4 @@
 from . import import_export_mcfg
 from . import import_export_p3d
-from . import import_export_rtm
-from . import props_action
 from . import props_material
 from . import props_object_mesh
